+++ src/main/java/com/butchercraft/transformation/serialization/TransformationDefinitionSerializer.java
@@ -1,0 +1,10 @@
+package com.butchercraft.transformation.serialization;
+
+import com.butchercraft.transformation.TransformationDefinition;
+
+/**
+ * Serializes transformation definitions into an external representation.
+ */
+public interface TransformationDefinitionSerializer<T> {
+    T serialize(TransformationDefinition definition);
+}